Create an instance of a fitting problem

Create an instance of a fitting problem

fitting_problem(ref_sample, controls = list(individual = individual_controls,
  group = group_controls), field_names, individual_controls, group_controls,
  prior_weights = NULL)

is.fitting_problem(x)

# S3 method for fitting_problem
format(x, ...)

# S3 method for fitting_problem
print(x, ...)

special_field_names(groupId, individualId, individualsPerGroup = NULL,
  count = NULL)

Arguments

ref_sample
The reference sample

controls
Control totals, by default initialized from the individual_controls and group_controls arguments

field_names
Names of special fields, construct using special_field_names()

individual_controls, group_controls
Control totals at individual and group level, given as a list of data frames where each data frame defines a control

prior_weights
Prior (or design) weights at group level; by default a vector of ones will be used, which corresponds to random sampling of groups

x
An object

...
Ignored.

groupId, individualId
Name of the column that defines the ID of the group or the individual

individualsPerGroup
Obsolete.

count
Name of control total column in control tables (use first numeric column in each control by default).

Value

An object of class fitting_problem, essentially a named list with the following components:

refSample

The reference sample, a data.frame.

controls

A named list with two components, individual and group. Each contains a list of controls as data.frames.

fieldNames

A named list with the names of special fields.